Tsiwela check

+++ ft_strlen.c
@@ -6,15 +6,15 @@
 /*   By: msefako <marvin@42.fr>                     +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2018/05/25 19:29:33 by msefako           #+#    #+#             */
-/*   Updated: 2018/05/25 19:29:36 by msefako          ###   ########.fr       */
+/*   Updated: 2018/06/15 16:53:27 by msefako          ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "libft.h"
 
-int		ft_strlen(const char *s)
+size_t		ft_strlen(const char *s)
 {
-	size_t i;
+	int i;
 
 	i = 0;
 	while (s[i])
